feat: set owner
