+++ src/api/clientSide/storeOwners.api.ts
@@ -20,31 +20,24 @@ const isStoreOwnerByUserId = async (userId: string) => {
   if (error) throw new Error(error.message);
 
   console.log(data);
-  // return data;
+  return data;
 };
-const isStoreOwnerByStoreIdAndUserId = async ({
-  storeId,
-  userId,
-}: {
-  storeId: string;
-  userId: string;
-}) => {
-  if (!storeId || !userId) return;
+const isStoreOwnerByStoreId = async ({ storeId }: { storeId: string }) => {
+  if (!storeId) return;
 
   const { error, data } = await supabase
     .from("storeOwners")
     .select()
-    .eq("storeId", storeId)
-    .eq("sponsorId", userId);
+    .eq("storeId", storeId);
   if (error) throw new Error(error.message);
 
-  console.log(data);
-  return data;
+  if (data.length === 0) return false;
+  return true;
 };
 
 const storeOwnersAPI = {
   insertStoreOwner,
-  isStoreOwnerByStoreIdAndUserId,
+  isStoreOwnerByStoreId,
   isStoreOwnerByUserId,
 };
 export default storeOwnersAPI;
