+++ src/components/Button/Button.tsx
@@ -39,6 +39,8 @@ const buttonVariant = cva("", {
       intent: "primary",
       className: "bg-opacity-40 font-bold",
     },
+    { outline: true, className: "bg-opacity-5" },
+    { outline: true, textIntent: "default", className: "!text-black" },
   ],
   defaultVariants: {
     outline: false,
